+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/CourseController.java
@@ -13,8 +13,10 @@ public class CourseController {
 
     SchoolService schoolService;
 
-    public CourseController(CourseService courseService) {
+    public CourseController(CourseService courseService, SchoolService schoolService){
+
         this.courseService = courseService;
+        this.schoolService = schoolService;
     }
 
     public Course createCourse(CourseDTO courseDTO, Long schoolId) {