+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/domain/School.java
@@ -29,7 +29,7 @@ public class School {
     private Center center;
 
     @OneToMany(mappedBy = "school")
-    @Column(name = "course_list")
+    @JsonManagedReference
     private List<Course> courseList;
 
     public School() {