+++ NIMIO2024_TFG_PolMorenoCastelar/changelog.md
@@ -26,3 +26,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - [NPP-16] 》SC8F - Get all courses of a school by ID
 - [NPP-17] 》SC9F - Edit a school by ID (entire data)
 - [NPP-18] 》SC10F - Delete a school by ID
+
+- [NPP-19] 》CO11F - Create a course that belongs to a school